requirements.txt
Flask
Flask-MySQLdb
python-dotenv
flask-restx
